Created API project

+++ Core/Calculators/CalculateState.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-using TPriceMeasure = System.Double;
+using TPriceMeasure = System.Decimal;
 using TProduct = System.String;
